+++ sockeye/arguments.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2017, 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not
 # use this file except in compliance with the License. A copy of the License
@@ -334,9 +334,8 @@ def add_bucketing_args(params):
 
     params.add_argument('--max-seq-len',
                         type=multiple_values(num_values=2, greater_or_equal=1),
-                        default=(100, 100),
-                        help='Maximum sequence length in tokens. Note that the target side will be extended by '
-                             'the <BOS> (beginning of sentence) token, increasing the effective target length. '
+                        default=(99, 99),
+                        help='Maximum sequence length in tokens.'
                              'Use "x:x" to specify separate values for src&tgt. Default: %(default)s.')
 
 
@@ -433,16 +432,16 @@ def add_model_parameters(params):
 
     model_params.add_argument('--encoder',
                               choices=C.ENCODERS,
-                              default=C.RNN_NAME,
+                              default=C.TRANSFORMER_TYPE,
                               help="Type of encoder. Default: %(default)s.")
     model_params.add_argument('--decoder',
                               choices=C.DECODERS,
-                              default=C.RNN_NAME,
+                              default=C.TRANSFORMER_TYPE,
                               help="Type of encoder. Default: %(default)s.")
 
     model_params.add_argument('--num-layers',
                               type=multiple_values(num_values=2, greater_or_equal=1),
-                              default=(1, 1),
+                              default=(6, 6),
                               help='Number of layers for encoder & decoder. '
                                    'Use "x:x" to specify separate values for encoder & decoder. Default: %(default)s.')
 
@@ -478,7 +477,7 @@ def add_model_parameters(params):
     # convolutional encoder/decoder arguments arguments
     model_params.add_argument('--cnn-kernel-width',
                               type=multiple_values(num_values=2, greater_or_equal=1, data_type=int),
-                              default=(3, 5),
+                              default=(3, 3),
                               help='Kernel width of the convolutional encoder and decoder. Default: %(default)s.')
     model_params.add_argument('--cnn-num-hidden',
                               type=int_greater_or_equal(1),
@@ -562,7 +561,7 @@ def add_model_parameters(params):
                               help='The type of positional embedding. Default: %(default)s.')
     model_params.add_argument('--transformer-preprocess',
                               type=multiple_values(num_values=2, greater_or_equal=None, data_type=str),
-                              default=('', ''),
+                              default=('n', 'n'),
                               help='Transformer preprocess sequence for encoder and decoder. Supports three types of '
                                    'operations: d=dropout, r=residual connection, n=layer normalization. You can '
                                    'combine in any order, for example: "ndr". '
@@ -572,7 +571,7 @@ def add_model_parameters(params):
                                    'Default: %(default)s.')
     model_params.add_argument('--transformer-postprocess',
                               type=multiple_values(num_values=2, greater_or_equal=None, data_type=str),
-                              default=('drn', 'drn'),
+                              default=('dr', 'dr'),
                               help='Transformer postprocess sequence for encoder and decoder. Supports three types of '
                                    'operations: d=dropout, r=residual connection, n=layer normalization. You can '
                                    'combine in any order, for example: "ndr". '
@@ -663,15 +662,19 @@ def add_training_args(params):
 
     train_params.add_argument('--batch-size', '-b',
                               type=int_greater_or_equal(1),
-                              default=64,
-                              help='Mini-batch size. Default: %(default)s.')
+                              default=4096,
+                              help='Mini-batch size. Note that depending on the batch-type this either refers to '
+                                   'words or sentences.'
+                                   'Sentence: each batch contains X sentences, number of words varies. '
+                                   'Word: each batch contains (approximately) X words, number of sentences varies. '
+                                   'Default: %(default)s.')
     train_params.add_argument("--batch-type",
                               type=str,
-                              default=C.BATCH_TYPE_SENTENCE,
+                              default=C.BATCH_TYPE_WORD,
                               choices=[C.BATCH_TYPE_SENTENCE, C.BATCH_TYPE_WORD],
-                              help="Sentence: each batch contains X sentences, number of words varies. Word: each batch"
-                                   " contains (approximately) X target words, number of sentences varies. "
-                                   "Default: %(default)s.")
+                              help="Sentence: each batch contains X sentences, number of words varies."
+                                   "Word: each batch contains (approximately) X target words, "
+                                   "number of sentences varies. Default: %(default)s.")
 
     train_params.add_argument('--fill-up',
                               type=str,
@@ -683,7 +686,7 @@ def add_training_args(params):
                               choices=[C.CROSS_ENTROPY],
                               help='Loss to optimize. Default: %(default)s.')
     train_params.add_argument('--label-smoothing',
-                              default=0.0,
+                              default=0.1,
                               type=float,
                               help='Smoothing constant for label smoothing. Default: %(default)s.')
     train_params.add_argument('--loss-normalization-type',
@@ -720,11 +723,11 @@ def add_training_args(params):
                               help='Maximum number of samples. Default: %(default)s.')
     train_params.add_argument(C.TRAIN_ARGS_CHECKPOINT_FREQUENCY,
                               type=int_greater_or_equal(1),
-                              default=1000,
+                              default=4000,
                               help='Checkpoint and evaluate every x updates/batches. Default: %(default)s.')
     train_params.add_argument('--max-num-checkpoint-not-improved',
                               type=int,
-                              default=8,
+                              default=32,
                               help='Maximum number of checkpoints the model is allowed to not improve in '
                                    '<optimized-metric> on validation data before training is stopped. '
                                    'Default: %(default)s.')
@@ -761,20 +764,20 @@ def add_training_args(params):
 
     train_params.add_argument('--rnn-decoder-hidden-dropout',
                               type=float,
-                              default=.0,
+                              default=.2,
                               help='Dropout probability for hidden state that combines the context with the '
                                    'RNN hidden state in the decoder. Default: %(default)s.')
     train_params.add_argument('--transformer-dropout-attention',
                               type=float,
-                              default=0.,
+                              default=0.1,
                               help='Dropout probability for multi-head attention. Default: %(default)s.')
     train_params.add_argument('--transformer-dropout-act',
                               type=float,
-                              default=0.,
+                              default=0.1,
                               help='Dropout probability before activation in feed-forward block. Default: %(default)s.')
     train_params.add_argument('--transformer-dropout-prepost',
                               type=float,
-                              default=0.,
+                              default=0.1,
                               help='Dropout probability for pre/postprocessing blocks. Default: %(default)s.')
     train_params.add_argument('--conv-embed-dropout',
                               type=float,
@@ -782,7 +785,7 @@ def add_training_args(params):
                               help="Dropout probability for ConvolutionalEmbeddingEncoder. Default: %(default)s.")
     train_params.add_argument('--cnn-hidden-dropout',
                               type=float,
-                              default=.0,
+                              default=.2,
                               help="Dropout probability for dropout between convolutional layers. Default: %(default)s.")
 
     train_params.add_argument('--optimizer',
@@ -818,13 +821,13 @@ def add_training_args(params):
                               help='Type of base weight initialization. Default: %(default)s.')
     train_params.add_argument('--weight-init-scale',
                               type=float,
-                              default=2.34,
+                              default=3.0,
                               help='Weight initialization scale. Applies to uniform (scale) and xavier (magnitude). '
                                    'Default: %(default)s.')
     train_params.add_argument('--weight-init-xavier-factor-type',
                               type=str,
-                              default='in',
-                              choices=['in', 'out', 'avg'],
+                              default=C.INIT_XAVIER_FACTOR_TYPE_AVG,
+                              choices=C.INIT_XAVIER_FACTOR_TYPES,
                               help='Xavier factor type. Default: %(default)s.')
     train_params.add_argument('--weight-init-xavier-rand-type',
                               type=str,
@@ -840,7 +843,7 @@ def add_training_args(params):
                                    'Default: %(default)s.')
     train_params.add_argument('--initial-learning-rate',
                               type=float,
-                              default=0.0003,
+                              default=0.0002,
                               help='Initial learning rate. Default: %(default)s.')
     train_params.add_argument('--weight-decay',
                               type=float,
@@ -857,7 +860,7 @@ def add_training_args(params):
                                    'Set to negative to disable. Default: %(default)s.')
     train_params.add_argument('--gradient-clipping-type',
                               choices=C.GRADIENT_CLIPPING_TYPES,
-                              default=C.GRADIENT_CLIPPING_TYPE_ABS,
+                              default=C.GRADIENT_CLIPPING_TYPE_NONE,
                               help='The type of gradient clipping. Default: %(default)s.')
 
     train_params.add_argument('--learning-rate-scheduler-type',
@@ -866,12 +869,12 @@ def add_training_args(params):
                               help='Learning rate scheduler type. Default: %(default)s.')
     train_params.add_argument('--learning-rate-reduce-factor',
                               type=float,
-                              default=0.5,
+                              default=0.7,
                               help="Factor to multiply learning rate with "
                                    "(for 'plateau-reduce' learning rate scheduler). Default: %(default)s.")
     train_params.add_argument('--learning-rate-reduce-num-not-improved',
                               type=int,
-                              default=3,
+                              default=8,
                               help="For 'plateau-reduce' learning rate scheduler. Adjust learning rate "
                                    "if <optimized-metric> did not improve for x checkpoints. Default: %(default)s.")
     train_params.add_argument('--learning-rate-schedule',
@@ -915,7 +918,7 @@ def add_training_args(params):
                               help="Names of parameters to fix at training time. Default: %(default)s.")
 
     train_params.add_argument(C.TRAIN_ARGS_MONITOR_BLEU,
-                              default=0,
+                              default=500,
                               type=int,
                               help='x>0: decode x sampled sentences from validation data and '
                                    'compute evaluation metrics. x==-1: use full validation data. Default: %(default)s.')
@@ -1087,6 +1090,11 @@ def add_inference_args(params):
                                type=float,
                                help='Beta factor for the length penalty used in beam search: '
                                     '(beta + len(Y))**alpha/(beta + 1)**alpha. Default: %(default)s')
+    decode_params.add_argument('--override-dtype',
+                               default=None,
+                               type=str,
+                               help='EXPERIMENTAL: may be changed or removed in future. Overrides training dtype of '
+                                    'encoders and decoders during inference. Default: %(default)s')
 
 def add_evaluate_args(params):
     eval_params = params.add_argument_group("Evaluate parameters")