+++ test/system/test_seq_copy_sys.py
@@ -16,9 +16,10 @@ import random
 
 import pytest
 
-logger = logging.getLogger(__name__)
-
+import sockeye.constants as C
 from test.common import tmp_digits_dataset, run_train_translate
+
+logger = logging.getLogger(__name__)
 
 _TRAIN_LINE_COUNT = 10000
 _DEV_LINE_COUNT = 100
@@ -36,73 +37,82 @@ _SEED_DEV_DATA = 17
 seed = random.randint(0, 1000)
 
 
+COMMON_TRAINING_PARAMS = " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001" \
+                         " --decode-and-evaluate 0 --label-smoothing 0.0" \
+                         " --optimized-metric perplexity --loss cross-entropy"
+
+
 @pytest.mark.parametrize("name, train_params, translate_params, use_prepared_data, perplexity_thresh, bleu_thresh", [
     ("Copy:lstm:lstm",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 16 --loss cross-entropy --optimized-metric perplexity"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001"
-     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --max-updates 4000 --weight-normalization"
-     " --gradient-clipping-type norm --gradient-clipping-threshold 10",
+     "--encoder rnn --decoder rnn "
+     " --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-attention-type mlp --rnn-attention-num-hidden 32"
+     " --batch-size 16 --batch-type sentence"
+     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --weight-normalization"
+     " --max-updates 4000"
+     " --gradient-clipping-type norm --gradient-clipping-threshold 10" + COMMON_TRAINING_PARAMS,
      "--beam-size 5 ",
      True,
      1.03,
      0.98),
     ("Copy:lstm:pruning",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 16 --loss cross-entropy --optimized-metric perplexity"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001"
+     "--encoder rnn --decoder rnn "
+     " --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
+     " --rnn-attention-num-hidden 32 --batch-size 16"
      " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --max-updates 4000 --weight-normalization"
-     " --gradient-clipping-type norm --gradient-clipping-threshold 10",
+     " --gradient-clipping-type norm --gradient-clipping-threshold 10" + COMMON_TRAINING_PARAMS,
      "--beam-size 5 --batch-size 2 --beam-prune 1",
      True,
      1.03,
      0.98),
     ("Copy:chunking",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 16 --loss cross-entropy --optimized-metric perplexity"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001"
-     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --max-updates 5000",
+     "--encoder rnn --decoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-attention-type mlp --rnn-attention-num-hidden 32"
+     " --batch-size 16 --batch-type sentence"
+     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0"
+     " --max-updates 5000" + COMMON_TRAINING_PARAMS,
      "--beam-size 5 --max-input-len 4",
      False,
      1.01,
      0.99),
     ("Copy:word-based-batching",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 80 --batch-type word --loss cross-entropy "
-     " --optimized-metric perplexity --max-updates 5000 --checkpoint-frequency 1000 --optimizer adam "
-     " --initial-learning-rate 0.001 --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --layer-normalization",
+     "--encoder rnn --decoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 "
+     " --rnn-attention-type mlp --rnn-attention-num-hidden 32 "
+     " --batch-size 80 --batch-type word "
+     " --max-updates 5000  "
+     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0 --layer-normalization" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      True,
      1.01,
      0.99),
     ("Copy:transformer:lstm",
-     "--encoder transformer --num-layers 2:1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
-     " --rnn-attention-type mhdot --rnn-attention-num-hidden 32 --batch-size 16 --rnn-attention-mhdot-heads 1"
-     " --loss cross-entropy --optimized-metric perplexity --max-updates 6000"
+     "--encoder transformer --decoder rnn --num-layers 2:1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-attention-type mhdot --rnn-attention-num-hidden 32 --rnn-attention-mhdot-heads 1"
+     " --max-updates 6000"
      " --transformer-attention-heads 4 --transformer-model-size 32"
      " --transformer-feed-forward-num-hidden 64 --transformer-activation-type gelu"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --batch-size 16 --batch-type sentence" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      False,
      1.01,
      0.99),
     ("Copy:lstm:transformer",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
-     " --decoder transformer --batch-size 16"
-     " --loss cross-entropy --optimized-metric perplexity --max-updates 3000"
+     "--encoder rnn --decoder transformer --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-decoder-hidden-dropout 0.0"
+     " --batch-size 16 --batch-type sentence"
+     " --max-updates 4000"
      " --transformer-attention-heads 4 --transformer-model-size 32"
-     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type swish1"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type swish1" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      True,
      1.01,
      0.98),
     ("Copy:transformer:transformer",
      "--encoder transformer --decoder transformer"
-     " --batch-size 16 --max-updates 4000"
+     " --max-updates 4000"
      " --num-layers 2 --transformer-attention-heads 4 --transformer-model-size 32"
      " --transformer-feed-forward-num-hidden 64 --num-embed 32"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --batch-size 16 --batch-type sentence" + COMMON_TRAINING_PARAMS,
      "--beam-size 1",
      False,
      1.02,
@@ -111,7 +121,7 @@ seed = random.randint(0, 1000)
      "--encoder cnn --decoder cnn "
      " --batch-size 16 --num-layers 3 --max-updates 4000"
      " --cnn-num-hidden 32 --cnn-positional-embedding-type fixed --cnn-project-qkv"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --batch-size 16 --batch-type sentence" + COMMON_TRAINING_PARAMS,
      "--beam-size 1",
      True,
      1.04,
@@ -133,7 +143,7 @@ def test_seq_copy(name, train_params, translate_params, use_prepared_data, perpl
                                                                     test_source_path=data['test_source'],
                                                                     test_target_path=data['test_target'],
                                                                     use_prepared_data=use_prepared_data,
-                                                                    max_seq_len=_LINE_MAX_LENGTH + 1,
+                                                                    max_seq_len=_LINE_MAX_LENGTH + C.SPACE_FOR_XOS,
                                                                     restrict_lexicon=True,
                                                                     work_dir=data['work_dir'],
                                                                     seed=seed)
@@ -147,69 +157,78 @@ def test_seq_copy(name, train_params, translate_params, use_prepared_data, perpl
 @pytest.mark.parametrize(
     "name, train_params, translate_params, use_prepared_data, use_source_factor, perplexity_thresh, bleu_thresh", [
     ("Sort:lstm:lstm",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 16 --loss cross-entropy --optimized-metric perplexity"
-     " --max-updates 7000 --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     "--encoder rnn --decoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-attention-type mlp"
+     " --rnn-attention-num-hidden 32"
+     " --max-updates 7000 "
+     " --batch-size 16 --batch-type sentence" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      True, False,
      1.03,
      0.97),
     ("Sort:word-based-batching",
-     "--encoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 --rnn-attention-type mlp"
-     " --rnn-attention-num-hidden 32 --batch-size 80 --batch-type word --loss cross-entropy"
-     " --optimized-metric perplexity --max-updates 6000 --checkpoint-frequency 1000 --optimizer adam "
-     " --initial-learning-rate 0.001 --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0",
+     "--encoder rnn --decoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32 "
+     " --rnn-attention-type mlp --rnn-attention-num-hidden 32 "
+     " --batch-size 80 --batch-type word"
+     " --max-updates 6000"
+     " --rnn-dropout-states 0.0:0.1 --embed-dropout 0.1:0.0" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      False, False,
      1.03,
      0.97),
     ("Sort:transformer:lstm",
-     "--encoder transformer --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
-     " --rnn-attention-type mhdot --rnn-attention-num-hidden 32 --batch-size 16 --rnn-attention-mhdot-heads 2"
-     " --loss cross-entropy --optimized-metric perplexity --max-updates 6000"
+     "--encoder transformer --decoder rnn --num-layers 1 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
+     " --rnn-attention-type mhdot --rnn-attention-num-hidden 32"
+     " --batch-size 16 --batch-type sentence"
+     " --rnn-attention-mhdot-heads 2"
+     " --max-updates 6000"
+     " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"
      " --transformer-attention-heads 4 --transformer-model-size 32"
-     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type gelu"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type gelu" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      True, False,
      1.03,
      0.97),
     ("Sort:lstm:transformer",
      "--encoder rnn --num-layers 1:2 --rnn-cell-type lstm --rnn-num-hidden 64 --num-embed 32"
-     " --decoder transformer --batch-size 16 --transformer-model-size 32"
-     " --loss cross-entropy --optimized-metric perplexity --max-updates 7000"
+     " --decoder transformer --transformer-model-size 32"
+     " --max-updates 7000"
      " --transformer-attention-heads 4"
      " --transformer-feed-forward-num-hidden 64 --transformer-activation-type swish1"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"
+     " --batch-size 16 --batch-type sentence" + COMMON_TRAINING_PARAMS,
      "--beam-size 5",
      False, False,
      1.03,
      0.97),
     ("Sort:transformer:transformer",
      "--encoder transformer --decoder transformer"
-     " --batch-size 16 --max-updates 6000"
-     " --num-layers 2 --transformer-attention-heads 4 --transformer-model-size 32 --num-embed 32"
-     " --transformer-feed-forward-num-hidden 64"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     " --batch-size 16 --batch-type sentence"
+     " --max-updates 6000"
+     " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 32 --num-embed 32"
+     " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"
+     " --transformer-feed-forward-num-hidden 64" + COMMON_TRAINING_PARAMS,
      "--beam-size 1",
      True, False,
      1.03,
      0.97),
     ("Sort:transformer_with_source_factor",
      "--encoder transformer --decoder transformer"
-     " --batch-size 16 --max-updates 6000"
+     " --batch-size 16 --batch-type sentence"
+     " --max-updates 6000"
      " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 32 --num-embed 32"
+     " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"
      " --transformer-feed-forward-num-hidden 64"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001 --source-factors-num-embed 2",
+     " --source-factors-num-embed 2" + COMMON_TRAINING_PARAMS,
      "--beam-size 1",
      True, True,
      1.03,
      0.96),
     ("Sort:cnn:cnn",
-     "--encoder cnn --decoder cnn "
-     " --batch-size 16 --num-layers 3 --max-updates 6000"
-     " --cnn-num-hidden 32 --cnn-positional-embedding-type fixed"
-     " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
+     "--encoder cnn --decoder cnn"
+     " --batch-size 16 --batch-type sentence"
+     " --max-updates 6000"
+     " --num-layers 3 --cnn-num-hidden 32 --cnn-positional-embedding-type fixed" + COMMON_TRAINING_PARAMS,
      "--beam-size 1",
      False, False,
      1.05,
@@ -239,7 +258,7 @@ def test_seq_sort(name, train_params, translate_params, use_prepared_data,
                                                                     test_source_factor_paths=data.get(
                                                                         'test_source_factors'),
                                                                     use_prepared_data=use_prepared_data,
-                                                                    max_seq_len=_LINE_MAX_LENGTH + 1,
+                                                                    max_seq_len=_LINE_MAX_LENGTH + C.SPACE_FOR_XOS,
                                                                     restrict_lexicon=True,
                                                                     work_dir=data['work_dir'],
                                                                     seed=seed)